# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Group:
    type: object
...
  models.Hike:
    properties:
      average_rating:
        type: number
      created_at:
        example: "2024-05-28T12:54:10.517438+02:00"
        type: string
      description:
        example: La rando de zinzin
        maxLength: 500
        type: string
      difficulty:
        enum:
        - Easy
        - Moderate
        - Hard
        example: Intermediate
        type: string
      duration:
        example: 3 hours
        type: string
      gpx_file:
        example: hike.gpx
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        example: 1
        type: integer
      image:
        example: hike_image.jpg
        type: string
      is_approved:
        default: false
        example: false
        type: boolean
      name:
        example: Montagne du destin
        maxLength: 100
        type: string
      organizer_id:
        example: 1
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/models.Subscription'
        type: array
      updated_at:
        example: "2024-05-28T12:54:10.517438+02:00"
        type: string
    required:
    - description
    - difficulty
    - duration
    - name
    - organizer_id
    type: object
...
  models.Subscription:
    properties:
      created_at:
        example: "2024-05-28T12:54:10.517438+02:00"
        type: string
      hike_id:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    required:
    - hike_id
    - user_id
    type: object
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /hikes:
    get:
      consumes:
      - application/json
      description: Get details of all hikes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Hike'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all hikes
      tags:
      - hikes
    post:
      consumes:
      - multipart/form-data
      description: Create a new hike with the input payload and optional image upload
      parameters:
      - description: Hike Name
        in: formData
        name: name
        required: true
        type: string
      - description: Hike Description
        in: formData
        name: description
        required: true
        type: string
      - description: Organizer ID
        in: formData
        name: organizer_id
        required: true
        type: integer
      - description: Difficulty
        in: formData
        name: difficulty
        required: true
        type: string
      - description: Duration
        in: formData
        name: duration
        required: true
        type: string
      - description: Is Approved
        in: formData
        name: is_approved
        type: boolean
      - description: Hike Image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hike'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new hike
      tags:
      - hikes
  /hikes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a hike by its ID
      parameters:
      - description: Hike ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete a hike by ID
      tags:
      - hikes
    get:
      consumes:
      - application/json
      description: Get details of a hike by its ID
      parameters:
      - description: Hike ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hike'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get a hike by ID
      tags:
      - hikes
    put:
      consumes:
      - multipart/form-data
      description: Update details of a hike by its ID and optional image upload
      parameters:
      - description: Hike ID
        in: path
        name: id
        required: true
        type: integer
      - description: Hike Name
        in: formData
        name: name
        required: true
        type: string
      - description: Hike Description
        in: formData
        name: description
        required: true
        type: string
      - description: Organizer ID
        in: formData
        name: organizer_id
        required: true
        type: integer
      - description: Difficulty
        in: formData
        name: difficulty
        required: true
        type: string
      - description: Duration
        in: formData
        name: duration
        required: true
        type: string
      - description: Is Approved
        in: formData
        name: is_approved
        type: boolean
      - description: Hike Image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hike'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update a hike by ID
      tags:
      - hikes
  /hikes/{id}/validate:
    patch:
      consumes:
      - application/json
      description: Validate a hike by its ID
      parameters:
      - description: Hike ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Validate a hike by ID
      tags:
      - hikes
  /hikes/notValidated:
    get:
      consumes:
      - application/json
      description: Get details of all hikes not validated
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Hike'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all hikes not validated
      tags:
      - hikes
  /hikes/subscribe:
    post:
      consumes:
      - application/json
      description: Post a subscription to a hike
      parameters:
      - description: Hike ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: formData
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Post a subscription to a hike
      tags:
      - hikes
...
swagger: "2.0"